rename edges subcommand to nodes
It's more useful to show information of nodes filtered by query.
